+++ crates/matrix-sdk-crypto/src/lib.rs
@@ -18,6 +18,7 @@
 
 #[cfg(feature = "backups_v1")]
 pub mod backups;
+pub mod dehydrated_devices;
 mod error;
 mod file_encryption;
 mod gossiping;